% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{scTenifoldNet.lite-package}
\alias{scTenifoldNet.lite}
\alias{scTenifoldNet.lite-package}
\title{scTenifoldNet.lite: Lightweight and Accelerated Gene Regulatory Network Comparison for Single-Cell Transcriptomics}
\description{
Fast and memory-efficient implementation of the scTenifoldNet workflow for constructing and comparing single-cell gene regulatory networks. The computational performance for principal component regression, tensor decomposition, and manifold alignment was optimized, while preserving the core analytical framework of the original scTenifoldNet method.
}
\author{
\strong{Maintainer}: Yuchen Li \email{ycli1995@outlook.com}

}
